.openclaw/workspace/projects/active/Treqe/backend/requirements.txt
# Backend Treqe - Dependencias principales
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# Base de datos
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Autenticación y seguridad
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Procesamiento asíncrono
celery==5.3.4
redis==5.0.1

# Validación de datos
pydantic==2.5.0
pydantic-settings==2.1.0

# Utilidades
python-dateutil==2.8.2
pytz==2023.3.post1
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.1

# Desarrollo
black==23.11.0
flake8==6.1.0
mypy==1.7.0